Refactor direct Chrome storage engine usage into a settings Facade
It simplifies changes to the way settings are loaded and saved.
So it was necessary to change the engine from 'sync' to 'local',
since sync has a QUOTA_BYTES_PER_ITEM = 8192,
i.e. it makes it impossible to save longer scripts
(without chopping strings into short parts
and putting them back together again)